refactor: rename label.py to text.py
